# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "패션 이커머스 StyleHub 백엔드 API. 로그인이 필요한 API 는 POST /api/v1/users/login 으로 받은 SESSION 쿠키로 인증한다.",
    "title" : "StyleHub API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/api/v1/orders" : {
      "get" : {
        "description" : "필요 역할: USER",
        "operationId" : "getMyOrders",
        "parameters" : [ {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CursorResponseOrderListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "내 주문 목록 조회(커서 페이징)",
        "tags" : [ "주문" ]
      },
      "post" : {
        "description" : "필요 역할: USER",
        "operationId" : "createOrder",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "주문 생성 — Idempotency-Key 로 재요청 시 중복 주문 방지",
        "tags" : [ "주문" ]
      }
    },
    "/api/v1/orders/{orderId}" : {
      "get" : {
        "description" : "필요 역할: USER",
        "operationId" : "getOrder",
        "parameters" : [ {
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "내 주문 상세 조회",
        "tags" : [ "주문" ]
      }
    },
...
    "/api/v1/users/me/addresses" : {
      "get" : {
        "description" : "필요 역할: USER",
        "operationId" : "getAddresses",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AddressResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "내 배송지 목록 조회",
        "tags" : [ "배송지" ]
      },
      "post" : {
        "description" : "필요 역할: USER",
        "operationId" : "registerAddress",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddressCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AddressResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "배송지 등록",
        "tags" : [ "배송지" ]
      }
    },
    "/api/v1/users/me/addresses/{addressId}" : {
      "delete" : {
        "description" : "필요 역할: USER",
        "operationId" : "deleteAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "배송지 삭제",
        "tags" : [ "배송지" ]
      }
    },
    "/api/v1/users/me/addresses/{addressId}/default" : {
      "patch" : {
        "description" : "필요 역할: USER",
        "operationId" : "changeDefaultAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AddressResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "기본 배송지 변경",
        "tags" : [ "배송지" ]
      }
    },
    "/api/v1/users/me/points" : {
      "get" : {
        "description" : "필요 역할: USER",
        "operationId" : "getMyPoints",
        "parameters" : [ {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyPointResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "summary" : "내 포인트 잔액·이력 조회",
        "tags" : [ "포인트" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AddressCreateRequest" : {
        "type" : "object",
        "properties" : {
          "detailAddress" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          },
          "label" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "phone" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^\\d{10,11}$"
          },
          "recipientName" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 2
          },
          "streetAddress" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          },
          "zipCode" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^\\d{5}$"
          }
        },
        "required" : [ "label", "phone", "recipientName", "streetAddress", "zipCode" ]
      },
      "AddressResponse" : {
        "type" : "object",
        "properties" : {
          "addressId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "detailAddress" : {
            "type" : "string"
          },
          "isDefault" : {
            "type" : "boolean"
          },
          "label" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "recipientName" : {
            "type" : "string"
          },
          "streetAddress" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          }
        }
      },
...
      "CursorResponseOrderListResponse" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderListResponse"
            }
          },
          "nextCursor" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "CursorResponsePointHistoryResponse" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PointHistoryResponse"
            }
          },
          "nextCursor" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "MyPointResponse" : {
        "type" : "object",
        "properties" : {
          "histories" : {
            "$ref" : "#/components/schemas/CursorResponsePointHistoryResponse"
          },
          "pointBalance" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "OrderCreateRequest" : {
        "type" : "object",
        "properties" : {
          "addressId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderDetailRequest"
            },
            "minItems" : 1
          },
          "usedPoint" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userCouponId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "addressId", "details" ]
      },
      "OrderDetailRequest" : {
        "type" : "object",
        "properties" : {
          "productOptionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "productOptionId", "quantity" ]
      },
      "OrderDetailResponse" : {
        "type" : "object",
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "orderDetailId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "productName" : {
            "type" : "string"
          },
          "productOptionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "string"
          },
          "storeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "storeName" : {
            "type" : "string"
          },
          "totalPrice" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unitPrice" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "OrderListResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "discountAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "earnedPoint" : {
            "type" : "integer",
            "format" : "int32"
          },
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orderStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "PAID", "PREPARING", "SHIPPING", "DELIVERED", "CANCELLED" ]
          },
          "pgOrderId" : {
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usedPoint" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "OrderResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderDetailResponse"
            }
          },
          "discountAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "earnedPoint" : {
            "type" : "integer",
            "format" : "int32"
          },
          "finalAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orderStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "PAID", "PREPARING", "SHIPPING", "DELIVERED", "CANCELLED" ]
          },
          "pgOrderId" : {
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usedPoint" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PointHistoryResponse" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "balanceSnapshot" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pointId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pointType" : {
            "type" : "string",
            "enum" : [ "EARN", "USE", "WELCOME", "DAILY_LOGIN" ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "session" : {
        "in" : "cookie",
        "name" : "SESSION",
        "type" : "apiKey"
      }
    }
  }
}